added LunchService interface
